# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ArchiveMaster API",
        "description": "Sistema de gestión documental empresarial completo",
        "contact": {
            "name": "ArchiveMaster Support",
            "email": "admin@archivemaster.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "ArchiveMaster API Server"
        }
    ],
    "paths": {
...
        "/api/statuses": {
            "get": {
                "tags": [
                    "Statuses"
                ],
                "summary": "Listar estados",
                "description": "Obtener lista de estados disponibles para la empresa",
                "operationId": "8bc31408b679bd52a317923523467108",
                "responses": {
                    "200": {
                        "description": "Lista de estados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "En Proceso"
                                                    },
                                                    "slug": {
                                                        "type": "string",
                                                        "example": "in_process"
                                                    },
                                                    "color": {
                                                        "type": "string",
                                                        "example": "blue"
                                                    },
                                                    "icon": {
                                                        "type": "string",
                                                        "example": "heroicon-o-clock"
                                                    },
                                                    "is_initial": {
                                                        "type": "boolean",
                                                        "example": false
                                                    },
                                                    "is_final": {
                                                        "type": "boolean",
                                                        "example": false
                                                    },
                                                    "active": {
                                                        "type": "boolean",
                                                        "example": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "timestamp": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/statuses/{id}": {
            "get": {
                "tags": [
                    "Statuses"
                ],
                "summary": "Obtener estado",
                "description": "Obtener detalles de un estado específico",
                "operationId": "700d24ee8a044d063a878f2c4555e35f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del estado",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del estado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "En Proceso"
                                                },
                                                "slug": {
                                                    "type": "string",
                                                    "example": "in_process"
                                                },
                                                "color": {
                                                    "type": "string",
                                                    "example": "blue"
                                                },
                                                "icon": {
                                                    "type": "string",
                                                    "example": "heroicon-o-clock"
                                                },
                                                "is_initial": {
                                                    "type": "boolean",
                                                    "example": false
                                                },
                                                "is_final": {
                                                    "type": "boolean",
                                                    "example": false
                                                },
                                                "active": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "timestamp": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Usar Bearer Token para autenticación",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}